+++ src/Core/ConfigTrait.php
@@ -48,7 +48,7 @@ trait ConfigTrait
      */
     public function config($key = null, $value = null)
     {
-        if (is_array($key) or  func_num_args() === 2) {
+        if (is_array($key) or func_num_args() === 2) {
             return $this->setConfig($key, $value);
         }
         return $this->getConfig($key);